cmake, onnxruntime dir doesn't work
