+++ olive/passes/onnx/custom_op_lib/csrc/custom_op_library.cc
@@ -1,19 +1,11 @@
 #include "custom_op_library.h"
 
-#define ORT_API_MANUAL_INIT
-#include "onnxruntime_cxx_api.h"
-#undef ORT_API_MANUAL_INIT
-
-#include <vector>
-#include <cmath>
-#include <mutex>
-#include <system_error>
-
+#include "core/session/onnxruntime_cxx_api.h"
 #include "core/common/common.h"
 #include "core/framework/ortdevice.h"
 #include "core/framework/ortmemoryinfo.h"
 #include "cuda/cuda_ops.h"
-#include "onnxruntime_lite_custom_op.h"
+#include "core/session/onnxruntime_lite_custom_op.h"
 
 static const char* c_OpDomain = "olive";
