+++ olive/passes/onnx/custom_op_lib/csrc/custom_op_library.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "onnxruntime_c_api.h"
+#include "core/session/onnxruntime_cxx_api.h"
 
 #ifdef __cplusplus
 extern "C" {
